Implemented the credentials of only user can add items for his account

+++ src/routes/ShoppingItemRoute.kt
@@ -1,10 +1,14 @@
 package com.omarahmed.routes
 
 import com.google.gson.Gson
+import com.omarahmed.data.models.ShoppingItem
+import com.omarahmed.data.models.User
 import com.omarahmed.data.requests.AddItemRequest
 import com.omarahmed.data.requests.UpdateItemRequest
 import com.omarahmed.data.responses.SimpleResponse
+import com.omarahmed.email
 import com.omarahmed.services.ShoppingItemService
+import com.omarahmed.services.UserService
 import com.omarahmed.util.Constants
 import com.omarahmed.util.Constants.BASE_URL
 import com.omarahmed.util.Constants.ITEMS_PICTURE_PATH
@@ -12,12 +16,14 @@ import com.omarahmed.util.QueryParams
 import com.omarahmed.util.saveFile
 import io.ktor.application.*
 import io.ktor.auth.*
+import io.ktor.auth.jwt.*
 import io.ktor.http.*
 import io.ktor.http.HttpStatusCode.Companion.BadRequest
 import io.ktor.http.HttpStatusCode.Companion.InternalServerError
 import io.ktor.http.HttpStatusCode.Companion.NoContent
 import io.ktor.http.HttpStatusCode.Companion.NotFound
 import io.ktor.http.HttpStatusCode.Companion.OK
+import io.ktor.http.HttpStatusCode.Companion.Unauthorized
 import io.ktor.http.content.*
 import io.ktor.request.*
 import io.ktor.response.*
@@ -27,7 +33,8 @@ import java.io.File
 import java.util.*
 
 fun Routing.addNewItemRoute(
-    shoppingItemService: ShoppingItemService
+    shoppingItemService: ShoppingItemService,
+    userService: UserService
 ) {
     authenticate {
         post("/api/items/new_item") {
@@ -38,7 +45,7 @@ fun Routing.addNewItemRoute(
             multiPart.forEachPart { partData ->
                 when (partData) {
                     is PartData.FormItem -> {
-                            itemName = partData.value
+                        itemName = partData.value
                     }
                     is PartData.FileItem -> {
                         val fileBytes = partData.streamProvider().readBytes()
@@ -55,16 +62,24 @@ fun Routing.addNewItemRoute(
 
             itemName?.let {
                 val itemImageUrl = "${BASE_URL}items_pictures/$fileName"
-                val addedItemAcknowledge = shoppingItemService.addNewItem(
-                    itemName = it,
-                    itemImageUrl = itemImageUrl
-                )
-                if (addedItemAcknowledge) {
-                    call.respond(OK, SimpleResponse<Unit>(true, "Successfully added new item!"))
+                val email = call.principal<JWTPrincipal>()?.email ?: ""
+                val userId = userService.getUserByEmail(email)?.id ?: ""
+                if (userService.doesEmailBelongToUserId(email, userId)) {
+                    val addedItemAcknowledge = shoppingItemService.addNewItem(
+                        userId = userId,
+                        itemName = it,
+                        itemImageUrl = itemImageUrl
+                    )
+                    if (addedItemAcknowledge) {
+                        call.respond(OK, SimpleResponse<Unit>(true, "Successfully added new item!"))
+                    } else {
+                        File("$ITEMS_PICTURE_PATH$fileName").delete()
+                        call.respond(InternalServerError)
+                    }
                 } else {
-                    File("$ITEMS_PICTURE_PATH$fileName").delete()
-                    call.respond(InternalServerError)
+                    call.respond(Unauthorized, "You are not who are")
                 }
+
             } ?: kotlin.run {
                 call.respond(BadRequest)
                 return@post
@@ -74,16 +89,20 @@ fun Routing.addNewItemRoute(
 
 }
 
-fun Route.getAllItemsRoute(shoppingItemService: ShoppingItemService) {
+fun Route.getAllItemsRoute(
+    shoppingItemService: ShoppingItemService,
+    userService: UserService
+) {
     authenticate {
         get("/api/items/get") {
             val page = call.parameters[QueryParams.PARAM_PAGE]?.toInt() ?: 0
             val pageSize = call.parameters[QueryParams.PARAM_PAGE_SIZE]?.toInt() ?: Constants.DEFAULT_PAGE_SIZE
-            val items = shoppingItemService.getItems(page, pageSize)
+            val email = call.principal<JWTPrincipal>()?.email ?: ""
+            val userId = userService.getUserByEmail(email)?.id ?: ""
+            val items = shoppingItemService.getItemsByUserId(userId,page, pageSize)
             call.respond(OK, items)
         }
     }
-
 }
 
 fun Route.updateItemRoute(shoppingItemService: ShoppingItemService) {
@@ -94,10 +113,10 @@ fun Route.updateItemRoute(shoppingItemService: ShoppingItemService) {
                 return@put
             }
             val updatedItem = call.receiveOrNull<UpdateItemRequest>()
-            if (updatedItem != null){
+            if (updatedItem != null) {
                 val isUpdateAcknowledge = shoppingItemService.updateItem(itemId, updatedItem)
                 if (isUpdateAcknowledge) {
-                    call.respond(OK, SimpleResponse<Unit>(true,"Successfully updated the item"))
+                    call.respond(OK, SimpleResponse<Unit>(true, "Successfully updated the item"))
                 } else {
                     call.respond(InternalServerError)
                 }
